requirements.txt
asgiref==3.2.10
backcall==0.2.0
colorama==0.4.3
debug==0.3.2
decorator==4.4.2
Django==3.0.8
django-crispy-forms==1.9.2
django-leaflet==0.27.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
Pygments==2.6.1
pytz==2020.1
see==1.4.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
